+++ src/components/cities/cities.connect.js
@@ -1,0 +1,7 @@
+import {setActiveCity} from "../../store/actions/city-actions";
+
+const mapDispatchToProps = (dispatch) => {
+  return {setActiveCity: (city) => dispatch(setActiveCity(city))};
+};
+
+export default mapDispatchToProps;
