+++ src/components/city-map/city-map.connect.js
@@ -1,0 +1,11 @@
+const mapStateToProps = ({HOTELS, CITY, MAP}) => {
+  return {
+    hotels: HOTELS.hotels,
+    nearby: HOTELS.nearby,
+    offer: HOTELS.offer,
+    activeCity: CITY.activeCity,
+    hover: MAP.hover,
+  };
+};
+
+export default mapStateToProps;
