+++ src/components/empty-main/empty-main.connect.js
@@ -1,0 +1,7 @@
+const mapStateToProps = ({CITY}) => {
+  return {
+    activeCity: CITY.activeCity,
+  };
+};
+
+export default mapStateToProps;
