+++ src/components/login/login.connect.js
@@ -1,0 +1,9 @@
+import {loginAction} from "../../store/actions/user-actions";
+
+const mapDispatchToProps = (dispatch) => {
+  return {
+    loginAction: (params) => dispatch(loginAction(params)),
+  };
+};
+
+export default mapDispatchToProps;
